Create MoveLog.cs
Added new autoload node, MoveLog, for storing and managing the log of moves made throughout the course of a game; Will eventually also be able to rewind/fast forward and replay games based on a given set of log entries.

Reorganized scripts into folders/namespaces as the quantity of scripts was getting high enough that it's worth doing.