Update Chessboard.cs
Updated handleMovementFinished to properly log castling moves; Instead of adding a second log for the rook moving after the king, update the Castle property of the king's move log entry; Only emit the PieceHasMoved signal when this is not the rook moving as a part of the castling move.

Game.handlePieceHasMoved can revert to assuming that it will only be called when selectedPiece is not null.