Update ChessPiece.cs
Added PieceNumber property for uniquely differentiating between pieces of the same team and type.

Update the actual movement tween operation to not require a BoardCell explicitly, only requiring that for the movement finished signal; Resetting the board for a new game will likely often require moving pieces from a graveyard back to the board.