Update MoveLogEntry.cs
The File property identifies when the moved piece occupies the same file as other of the same type and team; Likewise the Rank property identifies when the moved piece occupies the same rank as other of the same type and team. So when File && !Rank, display the rank; when !File && Rank, display the file.

Pawns are one of the exceptions:
- If a pawn is capturing a piece, display full origin.
- If a pawn is not capturing a piece, the origin can be assumed from the destination and team so omit the origin entirely.