Update Chessboard.cs
Removed the line clearing the move log in ResetPieces(); It is better to wait until all the pieces have finished moving so the log isn't immediately cluttered with all the moves from the reset animation.

Updated handleMovementFinished() so that ForceUpdate is called on the move log when castling is detected and the most recent entry is updated, ensuring the castle notation is displayed instead of just the king's move.