Update Game.cs
Added MoveLogView to the Game scene.
Updated handlePieceHasMoved() to handle the new GameStatus.Reseting status, waiting for all 32 pieces to finished returning to their starting positions before clearing the log and preparing the game state to begin playing again.

Updated handleStartNewGame() to set the game status to GameStatus.Reseting.

Updated toggleMainMenu() to hide the MoveLogView node when the main menu is visible and show it when the main menu is hidden.