Update MainMenu.cs
Added Credits node to the scene.

Added a button labeled Credits which displays the Credits node when clicked; Any keyboard key will hide the Credits node and show the main menu again when released.

Added readonly property CreditsVisible so that other nodes can easily detect if the credits node is visible or not.