Implement En Passant
Added a FirstMove property to MoveLogEntry.

Game when handlePieceHasMoved() triggers the reset condition, set the HasMoved property of every piece back to false so that En Passant can be detected in the new game.

CaptureLogic uses the MoveLog to determine if a pawn can perform En Passant move.

Chessboard checks for En Passant in handleMovementFinished() in order to capture the opposing pawn that was passed, since it will not be captured automatically by the normal capture mechanics.