Update Game.cs
Added call to CheckLogic.FilterKingMovesForCheck to handlePieceClicked, when the piece is a king, to only toggle the indicator on cells the king can actually move to.

Added call to board.DetectCheck to handleCapture when a king captures a piece to ensure the check status is removed if that piece was the only one causing the check status.