Update CheckLogic.cs
Added DetectCheckmateForTeam and DetectCheckmate methods for, as one might expect, detecting the checkmate state.

Updated FilterMovesToProtectKing to add cells to the empty move list instead of starting from the full list of potential moves and removing them.

Updated PredictCheck to rely on the BoardVector operator override.

Updated PredictCheckByCell to rely on the BoardVector operator override as well as validate that the destination is in-line between the king and the threatening piece.