Update ChessPiece.cs
Updated _InputEvent to have the parent cell emit the Clicked signal if the piece which was clicked is not currently listening for clicks; i.e. when capturing an opposing piece by clicking on the opposing piece.

Added _MouseEnter and _MouseExit overrides to call the parent's version of each respective method so that hovering over an opposing piece will properly highlight that cell, assuming it is a valid destination.