Create OptionsMenu.tscn
First draft of the options menu. There will eventually be sound, so adding an Audio tab. Also, the player should be able to remap the controls, so adding a Controls tab.
Basic script for the options menu enabling swapping between the different tabs.
Added stubs for the contents of the different options menu tabs.