Update OptionsMenu.tscn and OptionsMenu.cs
Added a button, Back, to the tab list for returning to the Main Menu.
Added a new signal, RequestHide, emitted when the back button is pressed.
Added public method ResetActiveTab to utilize showMenu to ensure the first menu is the active tab.
Updated showMenu so the parameter, id, has a default value of 0; Calling showMenu() with no parameter will always show the very first menu, as defined by the value order within the MenuId enumeration.