Update MainMenu.tscn and MainMenu.cs
Added an Options button to the Main Menu.
Added an OptionsMenu node to the MainMenu scene, hidden by default.
Updated the script so that Pressing the Options button displays the OptionsMenu and hides it again when the OptionsMenu.RequestHide signal is received.
Manually detach the RequestHide signal for completeness, even though the most likely scenario when this node would leave the tree is when the game itself is closing.