Update BoardCell.cs and ChessPiece.cs
Relying on the _InputEvent function precludes using keyboard input events. Instead, poll in _Processing and utilize _MouseEnter/_MouseExit to set the flag for when to poll.
Added readonly property ListeningForClicks to BoardCell so that ChessPiece only emits a signal on behalf of its occupied cell when that cell is marked to listen for clicks.