Update Game.tscn, TransparentBackgroundStyle.tres, GameUi.tscn, and 2 more files
Added GameUi scene and script; Currently displays the mapped inputs for the control actions along the bottom of the screen.
Added TransparentBackgroundStyle resource, self explanatory.
Added GameUi node to the Game scene and reordered the UI nodes.
Updated Game script to hide/show the GameUi node properly, as well as listen for the DeselectPressed signal.
Added _ExitTree override to Game script to detach all the custom signal handlers, for completeness.