Update Credits.tscn
Added a Back button to the Credits scene, which emits a BackPressed signal.
Updated MainMenu script to listen for and handle the Credits.BackPressed signal.
Copied the URL for all the LinkButton nodes into their Tooltip Text.
Updated the text for the 2d-chess-pack link from "2D Icons" to "2D Chess Icons"